+++ consai_examples/CMakeLists.txt
@@ -64,7 +64,7 @@ if(BUILD_TESTING)
   find_package(ament_cmake_pytest REQUIRED)
   ament_add_pytest_test(test_role_assignment tests/test_role_assignment.py)
   ament_add_pytest_test(test_robot_operator tests/test_robot_operator.py)
-  ament_add_pytest_test(test_field_observer tests/test_field_observer.py)
+  ament_add_pytest_test(test_pass_shoot_observer tests/test_pass_shoot_observer.py)
   ament_add_pytest_test(test_operation tests/test_operation.py)
   ament_add_pytest_test(test_detection_wrapper tests/test_detection_wrapper.py)
 endif()
