+++ consai_examples/consai_examples/decisions/attacker.py
@@ -17,7 +17,6 @@
 
 from consai_msgs.msg import State2D
 from decisions.decision_base import DecisionBase
-from field_observer import FieldObserver
 from operation import Operation
 from operation import TargetXY
 from operation import TargetTheta
