+++ consai_examples/consai_examples/decisions/center_back2.py
@@ -16,7 +16,6 @@
 # limitations under the License.
 
 from decisions.decision_base import DecisionBase
-from field_observer import FieldObserver
 from operation import Operation
 from operation import TargetXY
 from operation import TargetTheta
