+++ consai_examples/consai_examples/decisions/goalie.py
@@ -17,7 +17,6 @@
 
 from decisions.decision_base import DecisionBase
 
-from field_observer import FieldObserver
 from operation import Operation
 from operation import TargetXY
 from operation import TargetTheta
