+++ consai_examples/consai_examples/observer/ball_position_observer.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from consai_examples.observer.pos_vel import PosVel
 from consai_msgs.msg import State2D
 import math
